Fix gatling gun overheat problems
The overheating widget was oscillating between two states after having overheated once, and caused overheating when it looked like there was still enough room for firing.
The base game seems to correctly adjust the temperature, so it looks like there's no need to set it manually.
It is still required to set the overheating flag to trigger the overheat gun mesh state / animation